Fix java crash bug if students forget to log out.
SEVERE: java.time.format.DateTimeParseException: Text '-6:12:38' could not be parsed at index 0

A negative value is checked for and the error flag is set but this code was running irrespective of the error flag status.  Moved the crashing lines so they only evaluate if the error flag is not set.